fix(www): isolate landing pricing shader
